+++ kernel/string.cc
@@ -3,14 +3,14 @@
 #include "kernel.hh"
 
 void*
-memset(void *dst, int c, u32 n)
+memset(void *dst, int c, size_t n)
 {
   stosb(dst, c, n);
   return dst;
 }
 
 int
-memcmp(const void *v1, const void *v2, u32 n)
+memcmp(const void *v1, const void *v2, size_t n)
 {
   const u8 *s1, *s2;
   
@@ -26,7 +26,7 @@ memcmp(const void *v1, const void *v2, u32 n)
 }
 
 void*
-memmove(void *dst, const void *src, u32 n)
+memmove(void *dst, const void *src, size_t n)
 {
   const char *s;
   char *d;
@@ -47,13 +47,13 @@ memmove(void *dst, const void *src, u32 n)
 
 // memcpy exists to placate GCC.  Use memmove.
 void*
-memcpy(void *dst, const void *src, u32 n)
+memcpy(void *dst, const void *src, size_t n)
 {
   return memmove(dst, src, n);
 }
 
 int
-strncmp(const char *p, const char *q, u32 n)
+strncmp(const char *p, const char *q, size_t n)
 {
   while(n > 0 && *p && *p == *q)
     n--, p++, q++;
@@ -63,7 +63,7 @@ strncmp(const char *p, const char *q, u32 n)
 }
 
 char*
-strncpy(char *s, const char *t, u32 n)
+strncpy(char *s, const char *t, size_t n)
 {
   char *os;
   
@@ -77,7 +77,7 @@ strncpy(char *s, const char *t, u32 n)
 
 // Like strncpy but guaranteed to NUL-terminate.
 char*
-safestrcpy(char *s, const char *t, u32 n)
+safestrcpy(char *s, const char *t, size_t n)
 {
   char *os;
